organize directories, fix 3c4e data issue
