+++ src/components/ScrambleDisplayFrame.tsx
