+++ src/types/scramble-display.d.ts
